pyproject.toml
[project]
name = "prompt-tdd"
version = "0.1.0"
description = "Example of prompt engineering with test-driven development"
readme = "README.md"
requires-python = ">=3.12"
classifiers = ["Private :: Do Not Upload"]
dependencies = [
    "ollama>=0.4.7",
    "openai>=1.65.2",
    "pytest>=8.3.5",
    "pytest-testmon>=2.1.3",
    "pytest-watcher>=0.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = ["local: local-only (non-LLM) tests"]

[dependency-groups]
dev = ["factory-boy>=3.3.3", "isort>=6.0.1", "ruff>=0.9.9"]
